Yes, it works, and sounds.... COOL!
I got the basics of the Granular Synth working in which you can import a sample and then it plays through the sample breaking it into a number of splices and only playing a set percent of each splice. It loops through the splices and when it arrives at the end, it resarts for now.